requirements.txt
flake8==3.8.4
hiredis==1.1.0
mccabe==0.6.1
phone-iso3166==0.3.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
redis==3.5.3
redisbloom==0.4.0
redistimeseries==1.4.3
rmtest==0.7.0
six==1.15.0
